backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
bcrypt==4.1.2
openpyxl==3.1.2
reportlab==4.0.9
